Handle user input and call tap/swipe event handlers
